requirements.txt
azure-functions==1.0.0b4
azure-functions-worker==1.0.0b6
grpcio==1.20.1
grpcio-tools==1.20.1
protobuf==3.6.1
six==1.12.0
azureml-sdk==1.0.39
requests==2.20.1
scikit-image==0.14.2
scikit-learn==0.20.2
